-Extracted Auth Header logic in seperate component -Fixed styles and unclickable buttons problem in header -Changed folder structure -Fixed bugs and errors
